Use the `html_data` pageListParse for all.
It was only used in Webtoon side, but it starts on Manga too.
